requirements.txt
keras
librosa
scipy
matplotlib
numpy
pysoundfile
